+++ Core/UI/Lib/McuFont/decoder/mf_scaledfont.h
@@ -12,12 +12,12 @@ struct mf_scaledfont_s
     struct mf_font_s font;
 
     const struct mf_font_s *basefont;
-    uint8_t x_scale;
-    uint8_t y_scale;
+    float x_scale;
+    float y_scale;
 };
 
 MF_EXTERN void mf_scale_font(struct mf_scaledfont_s *newfont,
                              const struct mf_font_s *basefont,
-                             uint8_t x_scale, uint8_t y_scale);
+                             float x_scale, float y_scale);
 
 #endif
